figured out how people win
I changed the players to just 2 and now I'm going to brute force the
method that checks if someone won or not. (SO PAINFUL)
I've taken out a lot of the smaller methods that I was going to use,
but this also made it much easier to see the code that I'm working on.

Before I tried to add every cube int (the third thing for each piece)
and then use a recursive method to add them up in different groups of
4, but I couldn't figure out a way that would test every single
grouping.

After that I worked with zach on trying to find a way to check the
surrounding areas near the place that the player chose to see if there
are any pieces that would make him/her win. (E.g. just check the spaces
around the piece) but then I realized that since I'm doing mine 3d it
would take way too long since there are too many possibilities. (All
den if statements)
I then tried to find an equation that would search through the
different spaces, but then I again realized that the different spots
would have different specifications since subtracting out of the wrong
one would create an error.